+++ docs/components/Repl.vue
@@ -7,6 +7,10 @@ export default defineComponent({
   props: ['example', 'examplesImports', 'dependencies'],
   setup(props) {
     return () => {
+      if (typeof navigator === 'undefined') {
+        return null
+      }
+
       return h(Suspense, h(DocsRepl, props))
     }
   },