+++ packages/core/src/store/actions.ts
@@ -30,12 +30,15 @@ import {
   applyChanges,
   clamp,
   createAdditionChange,
+  createEdgeRemoveChange,
   createGraphNodes,
-  createRemoveChange,
+  createNodeRemoveChange,
   createSelectionChange,
-  getConnectedEdges,
+  getConnectedEdges as getConnectedEdgesBase,
   getDimensions,
   getHandleBounds,
+  getIncomers as getIncomersBase,
+  getOutgoers as getOutgoersBase,
   getOverlappingArea,
   getSelectionChanges,
   isDef,
@@ -53,7 +56,9 @@ export function useActions(
   hooksOn: any,
   state: State,
   getters: ComputedGetters,
+  // todo: change to a Set
   nodeIds: ComputedRef<string[]>,
+  // todo: change to a Set
   edgeIds: ComputedRef<string[]>,
 ): Actions {
   const viewportHelper = useViewport(state, getters)
@@ -62,6 +67,18 @@ export function useActions(
     const updateIds = ids ?? nodeIds.value ?? []
 
     state.hooks.updateNodeInternals.trigger(updateIds)
+  }
+
+  const getIncomers: Actions['getIncomers'] = (nodeOrId) => {
+    return getIncomersBase(nodeOrId, state.nodes, state.edges)
+  }
+
+  const getOutgoers: Actions['getOutgoers'] = (nodeOrId) => {
+    return getOutgoersBase(nodeOrId, state.nodes, state.edges)
+  }
+
+  const getConnectedEdges: Actions['getConnectedEdges'] = (nodesOrId) => {
+    return getConnectedEdgesBase(nodesOrId, state.edges)
   }
 
   const findNode: Actions['findNode'] = <T extends GraphNode = GraphNode>(id: string | undefined | null) => {
@@ -499,9 +516,11 @@ export function useActions(
     const edgeChanges: EdgeRemoveChange[] = []
 
     function createEdgeRemovalChanges(nodes: Node[]) {
-      const connections = getConnectedEdges(nodes, state.edges).filter((edge) => (isDef(edge.deletable) ? edge.deletable : true))
+      const connections = getConnectedEdges(nodes).filter((edge) => (isDef(edge.deletable) ? edge.deletable : true))
 
-      edgeChanges.push(...connections.map((connection) => createRemoveChange(connection.id)))
+      edgeChanges.push(
+        ...connections.map((connection) => createEdgeRemoveChange(connection.id, connection.source, connection.target)),
+      )
     }
 
     // recursively get all children and if the child is a parent, get those children as well until all nodes have been removed that are children of the current node
@@ -510,7 +529,7 @@ export function useActions(
 
       if (children.length) {
         const childIds = children.map((n) => n.id)
-        nodeChanges.push(...childIds.map((id) => createRemoveChange(id)))
+        nodeChanges.push(...childIds.map((id) => createNodeRemoveChange(id)))
 
         if (removeConnectedEdges) {
           createEdgeRemovalChanges(children)
@@ -533,7 +552,7 @@ export function useActions(
         return
       }
 
-      nodeChanges.push(createRemoveChange(currNode.id))
+      nodeChanges.push(createNodeRemoveChange(currNode.id))
 
       if (removeConnectedEdges) {
         createEdgeRemovalChanges([currNode])
@@ -570,7 +589,7 @@ export function useActions(
         return
       }
 
-      changes.push(createRemoveChange(typeof item === 'string' ? item : item.id))
+      changes.push(createEdgeRemoveChange(typeof item === 'string' ? item : item.id, currEdge.source, currEdge.target))
     })
 
     state.hooks.edgesChange.trigger(changes)
@@ -769,7 +788,7 @@ export function useActions(
 
     if (!state.d3Zoom) {
       until(() => state.d3Zoom)
-        .not.toBeUndefined()
+        .not.toBeNull()
         .then(setSkippedOptions)
     } else {
       setSkippedOptions()
@@ -898,6 +917,9 @@ export function useActions(
     setInteractive,
     setState,
     getIntersectingNodes,
+    getIncomers,
+    getOutgoers,
+    getConnectedEdges,
     isNodeIntersecting,
     panBy,
     fitView: (params) => viewportHelper.value.fitView(params),